% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempo.r
\name{tempo}
\alias{tempo}
\title{Set a tempo}
\usage{
tempo(bpm = 120)
}
\arguments{
\item{bpm}{A number specifying the tempo in beats per minute. Default is 120.}
}
\description{
\code{tempo} is a wrapper for Sys.time() that sets a tempo in bpm.
}
\examples{
# Set tempo at 60 beats per minute
tempo(60)

}
